+++ scripts/sync_protobuf/_utils.py
